Get_MC_Version() function
This proposed update will call the Get_MC_Version function to prompt the user to input the server version  they want to run and then download that version of paper.
The function uses wget return codes to  check the version exists before moving on. If the version doesn't exist the user will be informed and asked to enther the desired version again.
It would be super great to hear feedback on this if you don't want to include it, or to recieve contributor status/some credit on the github so I can put this contribution on my portfolio.
Thanks,
Fionn